backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
email-validator>=2.1.0
alembic>=1.13.0
